pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "portfolio-risk-engine"
version = "0.1.0"
description = "Standalone portfolio risk analytics engine"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pandas",
  "numpy",
  "statsmodels",
  "pyarrow",
  "cvxpy",
  "requests",
  "pyyaml",
]

[tool.hatch.build.targets.wheel]
packages = ["portfolio_risk_engine"]
